Move product attachment commands into dedicated namespace
