+++ src/Core/Domain/Product/Attachment/Command/SetAssociatedProductAttachmentsCommand.php
@@ -26,15 +26,14 @@
 
 declare(strict_types=1);
 
-namespace PrestaShop\PrestaShop\Core\Domain\Product\Command;
+namespace PrestaShop\PrestaShop\Core\Domain\Product\Attachment\Command;
 
 use PrestaShop\PrestaShop\Core\Domain\Attachment\ValueObject\AttachmentId;
+use PrestaShop\PrestaShop\Core\Domain\Product\Attachment\Command\RemoveAllAssociatedProductAttachmentsCommand;
 use PrestaShop\PrestaShop\Core\Domain\Product\ValueObject\ProductId;
 use RuntimeException;
 
 /**
- * @todo: should be moved to Product/Attachments ns along with command-related classes (in Adapter too) accordingly
- *
  * Replaces previous product attachments association with the provided one.
  */
 class SetAssociatedProductAttachmentsCommand
