+++ src/Core/Domain/Product/Attachment/CommandHandler/RemoveAllAssociatedProductAttachmentsHandlerInterface.php
@@ -24,9 +24,9 @@
  * @license   https://opensource.org/licenses/OSL-3.0 Open Software License (OSL 3.0)
  */
 
-namespace PrestaShop\PrestaShop\Core\Domain\Product\CommandHandler;
+namespace PrestaShop\PrestaShop\Core\Domain\Product\Attachment\CommandHandler;
 
-use PrestaShop\PrestaShop\Core\Domain\Product\Command\RemoveAllAssociatedProductAttachmentsCommand;
+use PrestaShop\PrestaShop\Core\Domain\Product\Attachment\Command\RemoveAllAssociatedProductAttachmentsCommand;
 
 /**
  * Defines contract to handle @see RemoveAllAssociatedProductAttachmentsCommand
