+++ src/PrestaShopBundle/Resources/config/services/adapter/product.yml
@@ -286,20 +286,20 @@ services:
               command: PrestaShop\PrestaShop\Core\Domain\Product\Command\BulkDeleteProductCommand
 
     prestashop.adapter.product.command_handler.set_associated_product_attachments_handler:
-        class: PrestaShop\PrestaShop\Adapter\Product\CommandHandler\SetAssociatedProductAttachmentsHandler
+        class: PrestaShop\PrestaShop\Adapter\Product\Attachment\CommandHandler\SetAssociatedProductAttachmentsHandler
         arguments:
             - '@prestashop.adapter.product.update.product_attachment_updater'
         tags:
             - name: tactician.handler
-              command: PrestaShop\PrestaShop\Core\Domain\Product\Command\SetAssociatedProductAttachmentsCommand
+              command: PrestaShop\PrestaShop\Core\Domain\Product\Attachment\Command\SetAssociatedProductAttachmentsCommand
 
     prestashop.adapter.product.command_handler.remove_all_associated_product_attachments_handler:
-        class: PrestaShop\PrestaShop\Adapter\Product\CommandHandler\RemoveAllAssociatedProductAttachmentsHandler
+        class: PrestaShop\PrestaShop\Adapter\Product\Attachment\CommandHandler\RemoveAllAssociatedProductAttachmentsHandler
         arguments:
             - '@prestashop.adapter.product.update.product_attachment_updater'
         tags:
             - name: tactician.handler
-              command: PrestaShop\PrestaShop\Core\Domain\Product\Command\RemoveAllAssociatedProductAttachmentsCommand
+              command: PrestaShop\PrestaShop\Core\Domain\Product\Attachment\Command\RemoveAllAssociatedProductAttachmentsCommand
 
     prestashop.adapter.product.validate.product_supplier_validator:
         class: PrestaShop\PrestaShop\Adapter\Product\Validate\ProductSupplierValidator
